+++ src/news/NewsCard.css
@@ -1,0 +1,4 @@
+.card {
+    display: inline-block;
+}
+
