+++ src/news/NewsForm.css
@@ -1,3 +1,4 @@
 form {
     max-width: 18rem;
+    display: inline-block;
 }
